+++ native/homm5-editor.c
@@ -82,6 +82,7 @@
 #include "qol/fix-payback.c"
 #include "qol/fix-dragon-form.c"
 #include "qol/fix-empowered-armageddon.c"
+#include "qol/fix-book-of-power.c"
 
 BOOL WINAPI DllMain(HINSTANCE self, DWORD reason, LPVOID reserved) {
   (void)reserved;
@@ -145,6 +146,7 @@ BOOL WINAPI DllMain(HINSTANCE self, DWORD reason, LPVOID reserved) {
   if (g_qol[QOL_PAYBACK_FIX]) install_payback_fix();
   if (g_qol[QOL_DRAGON_FORM_FIX]) install_dragon_form_fix();
   if (g_qol[QOL_EMPOWERED_ARMAGEDDON_FIX]) install_empowered_armageddon_fix();
+  if (g_qol[QOL_BOOK_OF_POWER_FIX]) install_book_of_power_fix();
   return TRUE;
 }